+++ zstl/test/org/zkoss/zktest/test2/F60/F60_ZK_951Test.scala
@@ -110,7 +110,7 @@ class F60_ZK_951Test extends ZTL4ScalaTestCase {
 			Change cols: 
 				<selectbox id="sbx3" model="${colsSize}">
 					<attribute name="onSelect">
-					int i = event.selectedObjects.iterator().next().intValue();
+					int i = event.selectedObjects.iterator().next();
 					if (biglist.frozenCols > i-1) {
 						self.selectedIndex = self.selectedIndex - 1;
 						alert("FrozenCols cannot be greater than Cols");
